Update patch set 1
Patch Set 1: Code-Review-1

(2 comments)

Hi Luca, Thanks for the doc review and patch!

One concern I have is one of the changes is affecting actual code. So either that plugin never worked in the first place and we and we had a typo. Or the upstream plugin actually expects the typo in their XML and this patch would then be breaking that support.

Can you double check the blame plugin and see which XML it is producing for that setting?

Patch-set: 1
Label: Code-Review=-1