Allow dashes in variable key names
It can be confusing when adding variables to templates and macros to
discover that '-' are not valid characters in the key name.

Expand variables as objects only if "obj:" is specified

It was possible to use an int variable as default: "x{var}x" but not as
default: "{var}" in a string parameter (which doesn't use
convert_mapping_to_xml). Also, somewhat unexpectedly, it used to be
possible to do default: "{dash-var}" as variables with dashes weren't
recognized by the CustomFormatter regex.

The documentation never claimed that "{var}" would expand as an object,
only "{obj:var}" was meant to, so let's make it behave so. It doesn't
break any tests, but it's possible some user code relies on it, so a
mention in release notes might be polite.

Change-Id: I8bf9ea4b725bb71b044fd5b6671f2fed46ceb1c9
Co-Authored-By: Tomas Janousek <tomas.janousek@gooddata.com>