Update patch set 1
Patch Set 1: Code-Review-1

I am against this move as is a defacto rule that variables can have only underlines and not dashes, because it will be impossible to parse grammar. I am not aware of any language that allows them.

I guess we would need a better "business case" to enable such thing.

Patch-set: 1
Label: Code-Review=-1