Update patch set 12
Patch Set 12:

Have not tested this change out. But there is lot of scope for improvement in the way logging and error message returned. I find that if the actual filename has typos|bad-code never prints the path/to/file-name as a part of the error output, which makes it hard to debug a large project with maybe 100s files. Is that something that can be addressed here?

Patch-set: 12
Reviewer: Gerrit User 10089 <10089@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Verified=0