Deprecate running jobs without --config-xml flag
Creation of nested folders is not possible in
``project-type: folder`` because the folder
name conflicts with the generated XML filename
and an ``OSError`` is thrown as a result. This
can be resolved by passing the ``--config-xml``
flag when running the jobs. This commit displays
a deprecation message asking the user to use
the ``--config-xml`` flag.

Change-Id: I6f883040ed3d9b3536ccc2fb305cdddcfde5d765
Story: #2001937
Task: #15050