Update patch set 1
Patch Set 1: Code-Review-1

Downvoting because the right place to document new option is inside the documentation, not on CRs.

Please fix this and I will be glad to make it a +2.

BTW, Are you sure that the unit tests do not need to be updated to test for this?

Patch-set: 1
Reviewer: Gerrit User 24162 <24162@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=-1