Update patch set 9
Patch Set 9:

There are a couple of problems we're running into here.

The first is that the dsl is generated as part of the root_xml, so it
doesn't get any templating applied to it (without this patch).

The second is that applying the full JJB templating to the dsl would be
a poor user experience, because it's a heavily curly-braced language,
which would require users to do a lot of {{'ing.

I'm not really sure how to resolve these, but changing the second would
probably require a major version bump (and would, I think, be bad for
users).

Patch-set: 9
Reviewer: Gerrit User 16703 <16703@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Verified=0