Update patch set 1
Patch Set 1: Code-Review-1

Hi Thierry, instead of creating an entirely new code review. It is better practice to update an existing code review. I'm not sure which one you're responding to but I would abandon this change and update the original patch that you pushed so that we can see the history of the code review.

Gerrit uses the "Change-Id" field in the footer of the commit message to determine if it needs to update an existing code review. So I would ensure that you change the Change-Id to match your previous patch using "git commit --amend"

Patch-set: 1
Reviewer: Gerrit User 13807 <13807@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=-1