Update patch set 2
Patch Set 2:

Talked to Darragh on IRC. I understand now, and am wondering why we even need this patch?

basepython by default is the python that is installed on your system. If you use python3 it will be python3, if you use python2 it will be python2.

Why do we need to force everyone to use python 3?

With that considered I think we should revert this patch and let the user choose their own python.

Patch-set: 2