Update patch set 4
Patch Set 4:

(6 comments)

I've pushed fixup changes for wrappers.py
but I;m not sure it is a good idea to leave minimal example without any secret because it does not make much sense as it will do nothing. But if that is acceptable I can update.

Patch-set: 4
Label: Verified=0