Update patch set 1
Patch Set 1:

My opinion is this is a documented feature so we should support it. If we really want to drop support then we need to flag this future as deprecated in the 2.x stream and remove it in the 3.x stream.

Vicky, I do not personally use the !!python/tuple notation though so if you could explain how you guys use it and what the benefit of it for you is that would probably help Sorin out.

Patch-set: 1