Update patch set 2
Patch Set 2: Code-Review-1

Ok so if we are following semvar that means we cannot remove this until we release v3.0.0 as you need to deprecate in one version (JJB 2.0) and remove in the next version (JJB 3.0).

If we are ready to jump to a 3.0 then we should probably do some sprint planning to figure out all of the features we want to break / remove.

Maybe we should gather all of the 3.0 related features into a 3.0 topic?

Patch-set: 2
Reviewer: Gerrit User 13807 <13807@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=-1