Update patch set 1
Patch Set 1:

> re: the import, I'm failing to understand why this is preferable.
 > Would you please tell me why we'd want to change that?

It's a suggestion from @Darragh. As I understand it in Python when you import a module it actually executes the module as code. So if you have 3 calls like:

 from module import func1
 from module import func2
 from module import func3

Module will execute code 3 times to import the function. By doing a single import we only execute code once.

Patch-set: 1