Update patch set 2
Patch Set 2: Code-Review-1

I thought we made the change to not explicitly require 'obj:' in order to be able to move away from requiring it in the future. This does speak to needing a better way to track decisions and things that should be scheduled for deprecation in the future better.

An example of a place that benefits is the top level disabled setting for jobs (and any other true/false based job setting), in that setting it from a template, it can be quite surprising that you must use 'obj:' as a prefix in order for the true or false boolean to be passed through correctly.

I wouldn't mind taking a stab at adjusting the regex to avoid the need to start requiring obj: again for Boolean's because that was a seriously annoying behaviour before.

Patch-set: 2
Label: Code-Review=-1