Update patch set 3
Patch Set 3: Code-Review-1

(1 comment)

I've added a test to this that shows the problem with not replacing booleans with objects by default.

AFAICS to resolve the underlying problem, we either need to alter all module settings dealing with integers to convert to strings when assigning to the final text, or switch all boolean settings to be able to perform a textual comparison of being passed 'true' or 'false' as a string in addition to the True or False objects in python.

Otherwise we're just exchanging one type of surprising behaviour for another type.

Patch-set: 3
Reviewer: Gerrit User 1054 <1054@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=-1