Update patch set 2
Patch Set 2:

> (1 comment)

I might defer IMHO here, the idea is to have a config option which is on par with CLI (--views-only | --jobs-only) so that users can set this in the config file and not having to provide this in the CLI every time. Adding a new option "update" looks less intuitive which changes the usability from the end users prespective. Maybe renaming add_views and add_jobs to views_only and jobs_only would is better?

Patch-set: 2