Update patch set 2
Patch Set 2:

> > (1 comment)
 > 
 > I might defer IMHO here, the idea is to have a config option which
 > is on par with CLI (--views-only | --jobs-only) so that users can
 > set this in the config file and not having to provide this in the
 > CLI every time. Adding a new option "update" looks less intuitive
 > which changes the usability from the end users prespective. Maybe
 > renaming add_views and add_jobs to views_only and jobs_only would
 > is better?

IMO having 2 options with invalid states is not a good user experience.


 Invalid: add_jobs=True && add_view=True
 Valid: add_jobs=True && add_view=False
 Valid: add_jobs=False && add_view=True
 Valid: add_jobs=False && add_view=False

This is not a good experience for a user of the software. It is better to have a single configuration when you have a states like this:

 update=job|views|all

Then programmatically we can wire up add_job and add_view the way it makes the most sense depending on the setting.

The reason this is not a problem on the CLI is because the user will only ever pass --jobs-only or --views-only on at a time. But in a config file where it's possible to have both configurations set it doesn't make sense to allow the user to set both of them and place the burden of knowing which options are incompatible with each other on the user.

Patch-set: 2