Update patch set 3
Patch Set 3:

yeah, that's what clued me in to the problem. Maybe subclassing or replacing part of the classes used by ElementTree to auto convert setting of 'text' to strings is the quickest way.

The alternative is to go through all of the code and ensure they are all set to string by default, which convert_mapping_to_xml already does, we just need to expand it to cover the remaining cases until they are converted.

I think the latter may not be too bad a change and certainly will fix all of our issues here while also allowing us to clean this code up by removing 'obj:' in the future.

Thoughts?

Patch-set: 3