Update patch set 1
Patch Set 1:

> @zxiiro Do you mean through email or here? Btw I'm wondering if the
 > issue can be seen in this link https://issues.sonatype.org/projects/NEXUS/issues/NEXUS-17832?filter=allopenissues

Yes I can see that link however that's not the same ticket you mentioned previously. NEXUS-17875 is set to "private" which means unless you're CC'd on the issue you cannot view it.

I don't know if you can do it but usually in Jira if you click on the number next to "Start watching" link on the right menu under "watchers" you should be able to add additional people to the CC of the ticket. If you add linuxfoundation-01 and linuxfoundation-02 accounts I will be able to view the ticket.

Patch-set: 1