Update patch set 2
Patch Set 2:

> @Daniel I saw you IRC your question regarding adding tests. Tests
 > are in the "tests" directory split between the different modules.
 > Each module has a "fixtures" directory.
 > 
 > Every YAML in the fixtures directory is a test and every XML with
 > the same name is a test output. You would write the test the same
 > way you would write the real world job config.
 > 
 > Hope this helps.

In order to reproduce this issue, the value for 'cancel-builds-on-update' has to come from a variable. The issue (as I understand it) is with the way deep_fromatter() formats the variables.

PyYAML loads the strings 'true/false' as python booleans, but then deep_formatter() applies the values with str.format() and all the booleans turn into strings.

Is there a way that I can use jjb variables in my test case?

Patch-set: 2
Label: Verified=0