Update patch set 1
Patch Set 1:

Hi Phil, thanks for the contribution!  This needs some tests adding; take a look at the existing !include-jinja2 tests for an idea of where to look.

The name of the directive isn't ideal, but that's my fault: when I contributed !include-jinja2 I named it incorrectly (as it is actually a parallel to !include-raw _not_ !include).

Patch-set: 1
Label: Verified=0