Update patch set 4
Patch Set 4:

> Looks like it's breaking 2 unit tests. Logs here:
 > 
 > https://zuul.opendev.org/t/openstack/build/42442641586c4e3a922c4143a840dc6d/log/job-output.txt#1652-1696
 > 
 > Both with the message "KeyError: 'folder'" when it's trying to
 > access data['folder']. I'm not too familiar with this part of the
 > code unfortunately.

Thanks! I figured it out and fixed them.

Is there a better way to identify what the properties are affecting?  The `_use_folder_perms` feels very hacky to me, but the XML Element doesn't have a way to inspect up the tree, and the current code flow would need an overhaul to pass the top-level reference down.

Patch-set: 4