Update patch set 7
Patch Set 7: Code-Review-1

(2 comments)

Even though it's always useful to test explicitly setting the same as the default, it seems like it might be more important to test that not setting the value results in the expected and that setting it to something other than False results in the expected output as well.

I see the test values have fluctuated a bit during the different revisions, but I think might want to go back to the original values or add a 3rd test case if there is concern about the user setting 'false' explicitly not working as expected.

Particular as some of the output generated before was missing the boolean appearing in the XML, which based on the text description, one would expect not setting this option should still result in 'false' appearing in the XML output rather than an empty tag. If not there is something else missing that needs to be set to force the tag to contain the value using the mapping code.

Patch-set: 7
Label: Code-Review=-1