Update patch set 14
Patch Set 14:

> thanks Thanh,
 > I wonder which of the two is better: 1. merge the change just
 > before the new pypi release, 2. merge it now, and revert/fix
 > potential issues ppl that use the master branch may find, and
 > (potentially) have fixes before the new pypi release?

We typically merge this kind of change as adding support for new XML to plugins is usually pretty low risk. Either the plugin supports the new XML or completely ignores it if it doesn't.

Our usual process though is to have 2 reviewers review a patch before it is merged. Sometimes we have exceptions if folks are not quick to respond.

Patch-set: 14