Update patch set 4
Patch Set 4: Code-Review-1

(3 comments)

A few suggestions I'd like to see fixed but otherwise the patch looks good.

(I saw you abandon another patch, in the future if you're recreating the patch remember to set the "Change-Id" in the footer of your commit message, Gerrit uses that to identify which existing review to patch so that you're not alwys creating brand new reviews.

Patch-set: 4
Label: Code-Review=-1