Update patch set 7
Patch Set 7:

> Thanh, do you believe that somebody else going to review it? Or
 > shall we merge it as it is? It would be really great to have this
 > available in the next release.

Not sure but I am also the only person who performs the releases and I typically merge all patches that have +2's when I make a release, either way it will be in the next release regardless of if no one else steps up to review.

I'm currently on vacation and do not have plans to create a new release until I return so there is still time for folks.

If you want to help out there are other patches in JJB which require reviewing too which I've already +2'd it would be very useful to have more eyes on patches rather than just myself. If more folks from the community steps up to help (even if they can only +1 the patch) we can merge more patches more quickly.

JJB is something I am maintaining in my spare time and I do worry that if more folks don't step up to help out the longevity of this project may be at risk. So if this project is useful to you in anyway please do consider helping out with some reviews and encouraging others to do the same.

Patch-set: 7