Update patch set 1
Patch Set 1:

Not entirely sure what's going on myself but it appears to be failing on Python 2 and passing on Python 3. I cannot commit to investigating this further at the moment but I'll look into this once I have some more free cycles if no one else get's to it first.

Patch-set: 1