Introduce disable-host-key-checking for builder ansible-playbook
Implemented new parameter "disable-host-key-checking" to make the
"anbile-playbook" builder compatible with version >=1.0 of the ansible plugin.
The parameter defaults to "false" for security reasons (as the plugin does).
The old parameter "host-key-checking" is still written to XML to keep code
working in a setup with plugin version <1.0. This has no impact on newer setups
because the plugin with version >=1.0 ignores the old parameter value.
For same security reasons the default value of old parameter
"host-key-checking" was changed to "true" to "fix" older setups which have the
parameter not defined.

Change-Id: Icdbf58bea0b583f6d5fe11a497b3b03eecef3c5d
Task: 39789
Story: 2007678