Update patch set 2
Patch Set 2:

Looks like the failure unrelated to this patch is due to http://viewgit.fealdia.org/ now being a broken link. Not sure what's going on there, if this doesn't fix itself we'll have to remove that link from the SCM plugin pointing that that via another patch.

Patch-set: 2
Label: Verified=0