Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> Both parameters do not have to be user configurable:

So what you are saying is that these new parameters should not be documented because a person using this plugin would never want to set either of the 2 parameters being configured here?

Is that the assumption we are taking on these items?

Patch-set: 1
Reviewer: Gerrit User 13807 <13807@4a232e18-c5a9-48ee-94c0-e04e7cca6543>